+++ gaussian.cpp
@@ -1,0 +1,52 @@
+/**
+ * @short   gaussion
+ * @file    gaussian.cpp
+ * @author  Eric Jambo
+ *
+ * This file contains the prototypes and a short description of all the functions used
+ * to generate a guassian function and its derivative using total time and time steps.
+ */
+
+#include <iostream>
+#include <string>
+#include <fstream>
+#include <sstream>
+#include <stdlib.h>
+#include <cstdlib>
+#include <cmath>
+#include "gaussian.h"
+
+using namespace std;
+
+// defining the value of PI.
+
+#define PI 3.14159265
+
+double gauss_func (double t[], double t_der[])
+{
+
+    cout << "Running: gaussian_function\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function calculates the derivative of the input waveform using total
+ *                    time and time steps
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double der_wavf_func (double t[], double t_der[])
+{
+
+    cout << "Running: derivative_waveform_function\n";
+    cout << endl;
+
+    return 0;
+}
+
